Title: 특정 수와 근접한 합, Time: -1ms, Memory: -1MB, Status: Compile Error - CodeTree

+++ 241105/특정 수와 근접한 합/sum-close-to-particular-number.cpp
@@ -7,7 +7,7 @@ int main() {
     int* arr;
     scanf("%d %d", &N, &S);
 
-    arr = new int(N);
+    arr = new short(N);
 
     for(int i=0; i<N; i++)
     {
